requirements.txt
Flask
requests
python-dotenv

torch
torchvision
torchaudio

jupyterlab
ipywidgets